Move bm interface to common
