+++ extension/src/BookmarksPanel/mapper/index.ts
@@ -1,4 +1,8 @@
-import { ContextBookmark, IBookmarksObj, IFolderMetaData } from '../interfaces';
+import {
+  ContextBookmark,
+  IBookmarksObj,
+  IFolderMetaData,
+} from '@common/components/Bookmarks/interfaces';
 
 export const bookmarksMapper: (
   kvp: [string, IFolderMetaData],
