+++ extension/src/BookmarksPanel/utils/index.ts
@@ -7,7 +7,7 @@ import {
   IBookmark,
   IBookmarksObj,
   ISelectedBookmarks,
-} from '../interfaces';
+} from '@common/components/Bookmarks/interfaces';
 
 export const getFaviconUrl = (url: string) =>
   `https://www.google.com/s2/favicons?sz=64&domain_url=${new URL(url).origin}`;
